examples/mnist/image/requirements.txt
numpy==1.14.3
Pillow==5.2.0
tensorflow==1.9.0
hs==0.1.3